lxd/response: Split SmartError into LXD/non-LXD
