+++ lxd/response/response.go
@@ -2,7 +2,6 @@ package response
 
 import (
 	"bytes"
-	"database/sql"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -11,14 +10,10 @@ import (
 	"os"
 	"time"
 
-	"github.com/canonical/go-dqlite/driver"
 	lxd "github.com/lxc/lxd/client"
-	"github.com/lxc/lxd/lxd/db"
 	"github.com/lxc/lxd/lxd/util"
 	"github.com/lxc/lxd/shared"
 	"github.com/lxc/lxd/shared/api"
-	"github.com/mattn/go-sqlite3"
-	"github.com/pkg/errors"
 )
 
 var debug bool
@@ -192,38 +187,6 @@ func Unavailable(err error) Response {
 		message = err.Error()
 	}
 	return &errorResponse{http.StatusServiceUnavailable, message}
-}
-
-// SmartError returns the right error message based on err.
-func SmartError(err error) Response {
-	if err == nil {
-		return EmptySyncResponse
-	}
-
-	switch errors.Cause(err) {
-	case os.ErrNotExist, sql.ErrNoRows, db.ErrNoSuchObject:
-		if errors.Cause(err) != err {
-			return NotFound(err)
-		}
-
-		return NotFound(nil)
-	case os.ErrPermission:
-		if errors.Cause(err) != err {
-			return Forbidden(err)
-		}
-
-		return Forbidden(nil)
-	case db.ErrAlreadyDefined, sqlite3.ErrConstraintUnique:
-		if errors.Cause(err) != err {
-			return Conflict(err)
-		}
-
-		return Conflict(nil)
-	case driver.ErrNoAvailableLeader:
-		return Unavailable(err)
-	default:
-		return InternalError(err)
-	}
 }
 
 func (r *errorResponse) String() string {
